moved ChipController into its own module
