+++ src/chip_controller/chip/debug/rom.rs
