+++ src/chip_controller/chip/debug/sprite.rs
