+++ src/chip_controller/chip/display.rs
