+++ src/chip_controller/chip/instruction.rs
