+++ src/chip_controller/chip/mod.rs
